feat: update to ntopng v2 API
As of ntopng 5.0 the v1 REST API is deprecated and it says that it will
be completely removed in the next stable release.

See: https://github.com/ntop/ntopng/releases/tag/5.0

After looking into it, we don't appear to be using any fields that were
removed in the v2 version of the API. So this can be processed mostly in
place.

There are new fields though that we might be able to take advantage of
at a later date.